requirements.txt
beautifulsoup4==4.6.1
cachetools==2.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==2.0.3
django-tables2==1.21.2
Flask==1.0.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.6
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.4
mail-parser==3.4.1
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.2
oauth2client==4.1.2
oauthlib==2.0.7
pandas==0.22.0
pbr==3.1.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
PySocks==1.6.8
python-dateutil==2.7.2
pytz==2018.3
requests==2.19.1
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.16.0
six==1.11.0
stevedore==1.28.0
tweepy==3.6.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
